requirements.txt
pynacl>=1.5.0
zeroconf>=0.40.0
textual>=0.25.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
